+++ core/src/main/java/com/graphhopper/routing/util/parsers/FootAverageSpeedParser.java
@@ -20,7 +20,7 @@ public class FootAverageSpeedParser extends AbstractAverageSpeedParser implement
                 lookup.getDecimalEncodedValue(FerrySpeed.KEY));
     }
 
-    protected FootAverageSpeedParser(DecimalEncodedValue speedEnc, DecimalEncodedValue ferrySpeedEnc) {
+    public FootAverageSpeedParser(DecimalEncodedValue speedEnc, DecimalEncodedValue ferrySpeedEnc) {
         super(speedEnc, ferrySpeedEnc);
 
         routeMap.put(INTERNATIONAL, UNCHANGED.getValue());